問題生成完了

+++ api/src/main/java/com/example/api/entity/Artist.java
@@ -24,7 +24,7 @@ public class Artist {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "artist_id", nullable = false)
-    private Integer artistId;
+    private Long artistId;
 
     /**
      * アーティスト名
